rename library from TVout to TVout_ve
All relevant file and folder names were changed to suit the new library
name and all references in any relevant files were also changed.
