+++ TVout_ve/examples/EdgeDetection/edge_templates.h
