+++ TVout_ve/spec/hardware_setup.h
